chore: refactor WrappingHStack
